chore(skills): split server-owned interception tests under the 1200-line cap
Task 3 / executeServerOwned coverage moved to
tests/unit/skills-interception-server-owned.test.ts so the original file
stays at the tip size. Vietnamese locale now has the
SERVER_OWNED_TOOL_LOOP_ENABLED keys. New unit files on this branch are
listed in stryker tap.testFiles. chat-pipeline.test.ts is frozen at 1648
(+4) because the upstream-500 case now turns request retries off so the
provider pipeline cannot double-count fetch.
