+++ stryker.conf.json
@@ -237,6 +237,7 @@
       "tests/unit/cursor-renewal.test.ts",
       "tests/unit/custom-model-target-format.test.ts",
       "tests/unit/db-reset-module-state.test.ts",
+      "tests/unit/db-server-tool-executions-migration.test.ts",
       "tests/unit/db/stats-dbstat-optional.test.ts",
       "tests/unit/ddg-circuit-breaker-null-content-6999-7000.test.ts",
       "tests/unit/domain-persistence.test.ts",
@@ -253,6 +254,7 @@
       "tests/unit/executor-devin-cli-agentic-acp.test.ts",
       "tests/unit/executor-web-cookie-sweep.test.ts",
       "tests/unit/false-terminal-401-quota.test.ts",
+      "tests/unit/follow-up-transcript.test.ts",
       "tests/unit/format-provider-error-cause.test.ts",
       "tests/unit/forwarded-header-budget.test.ts",
       "tests/unit/fusion-vision-panel-3378.test.ts",
@@ -297,6 +299,7 @@
       "tests/unit/noauth-autocombo-lockout-7623.test.ts",
       "tests/unit/ollama-404-model-lockout-11071.test.ts",
       "tests/unit/non-streaming-client-translate.test.ts",
+      "tests/unit/non-streaming-provider-leg.test.ts",
       "tests/unit/non-streaming-sse-terminal-typescan-4459.test.ts",
       "tests/unit/nvidia-410-model-scope.test.ts",
       "tests/unit/nvidia-passthrough-models-6773.test.ts",
@@ -373,6 +376,10 @@
       "tests/unit/serial/combo-quota-share-cooldown-wait-timing.test.ts",
       "tests/unit/serial/combo-strategy-fallbacks-half-open-timing.test.ts",
       "tests/unit/serial/provider-health-autopilot.test.ts",
+      "tests/unit/server-owned-tool-loop-flag.test.ts",
+      "tests/unit/server-owned-tool-loop.test.ts",
+      "tests/unit/skill-execution-fence.test.ts",
+      "tests/unit/skills-interception-server-owned.test.ts",
       "tests/unit/service-combo-metrics.test.ts",
       "tests/unit/service-provider-plugin-registry.test.ts",
       "tests/unit/services-branch-hardening.test.ts",
@@ -386,6 +393,7 @@
       "tests/unit/sse-auth-exclusive-leases.test.ts",
       "tests/unit/sse-auth-resource-404.test.ts",
       "tests/unit/sse-auth.test.ts",
+      "tests/unit/stable-json.test.ts",
       "tests/unit/stream-early-eof-breaker.test.ts",
       "tests/unit/stream-readiness.test.ts",
       "tests/unit/strict-random-deck.test.ts",
@@ -396,6 +404,7 @@
       "tests/unit/thundering-herd.test.ts",
       "tests/unit/token-refresh-race-comprehensive.test.ts",
       "tests/unit/token-refresh-service.test.ts",
+      "tests/unit/tool-loop-usage.test.ts",
       "tests/unit/tools-filter-anthropic-format.test.ts",
       "tests/unit/tproxy-route.test.ts",
       "tests/unit/trae-publiccred.test.ts",
